+++ apps/api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lexion-rte/api
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @lexion-rte/core@0.1.3
+  - @lexion-rte/extensions@0.1.3
+  - @lexion-rte/tools@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes