+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @lexion-rte/playground
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @lexion-rte/core@0.1.3
+  - @lexion-rte/extensions@0.1.3
+  - @lexion-rte/web@0.1.3
+  - @lexion-rte/vue@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes